move to show_elfinder to utils
